requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.1.4
pdfplumber==0.10.3
PyMuPDF==1.23.9
pytesseract==0.3.10
Pillow==10.1.0
openai==1.3.8
google-generativeai==0.3.2
azure-openai==1.0.0b1
python-dotenv==1.0.0
structlog==23.2.0
email-validator==2.1.0
